requirements.txt
wheel
numpy
pandas
tensorflow-gpu
Pillow
jupyter
matplotlib
